requirements.txt
Flask
deep-translator
python-docx
fpdf
PyPDF2
Flask-Cors